+++ backend/src/main/java/com/tibame/app_generator/dto/analysis/FieldDTO.java
@@ -1,0 +1,18 @@
+package com.tibame.app_generator.dto.analysis;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class FieldDTO {
+    private String name;
+    private String type;
+    private List<String> modifiers;
+}